Catches up with 300 develop commits. The only overlap with this branch is the
shared mspecs, all three of which develop moved on:

- knx-master-data.mspec gained vendor IDs -> KnxManufacturer.cs (the one file
  git could not auto-merge) regenerated with the .NET generator.
- s7.mspec: COTPPacket length arg uint 16 -> uint 32, an S7Payload-optional
  guard, DATE_AND_TIME dayOfWeek -> a manual field. S7 model regenerated; all
  80 S7 tests still pass (the generator's DATE_AND_TIME handler already reads
  dayOfWeek as a plain nibble and recomputes it on serialize).
- modbus.mspec: the DataItem dataIo was redesigned to one value at its natural
  width - register alignment, arrays and byte order are the driver's job now -
  and STRING became `vstring 'stringLength * 8'`. Taught the generator's dataIo
  path to emit a vstring read/write/length from its run-time length expression
  (it already did for regular fields), regenerated, and rewrote
  ModbusDataItemRoundTripTests for the new single-value signature. The Modbus
  driver does not use DataItem, so no driver change.

All three models regenerate deterministically (CI drift check passes).
413 tests, 0 warnings.

+++ plc4go/protocols/abeth/readwrite/model/CIPEncapsulationConnectionRequest.go
@@ -185,13 +185,13 @@ func (m *_CIPEncapsulationConnectionRequest) GetPlx4xTypeName() string {
 	return "CIPEncapsulationConnectionRequest"
 }
 
-func (m *_CIPEncapsulationConnectionRequest) GetLengthInBits(ctx context.Context) uint16 {
-	lengthInBits := uint16(m.CIPEncapsulationPacketContract.(*_CIPEncapsulationPacket).getLengthInBits(ctx))
+func (m *_CIPEncapsulationConnectionRequest) GetLengthInBits(ctx context.Context) uint64 {
+	lengthInBits := uint64(m.CIPEncapsulationPacketContract.(*_CIPEncapsulationPacket).getLengthInBits(ctx))
 
 	return lengthInBits
 }
 
-func (m *_CIPEncapsulationConnectionRequest) GetLengthInBytes(ctx context.Context) uint16 {
+func (m *_CIPEncapsulationConnectionRequest) GetLengthInBytes(ctx context.Context) uint64 {
 	return m.GetLengthInBits(ctx) / 8
 }
