+++ plc4go/protocols/ads/discovery/readwrite/model/AdsDiscoveryBlockStatus.go
@@ -227,8 +227,8 @@ func (m *_AdsDiscoveryBlockStatus) GetPlx4xTypeName() string {
 	return "AdsDiscoveryBlockStatus"
 }
 
-func (m *_AdsDiscoveryBlockStatus) GetLengthInBits(ctx context.Context) uint16 {
-	lengthInBits := uint16(m.AdsDiscoveryBlockContract.(*_AdsDiscoveryBlock).getLengthInBits(ctx))
+func (m *_AdsDiscoveryBlockStatus) GetLengthInBits(ctx context.Context) uint64 {
+	lengthInBits := uint64(m.AdsDiscoveryBlockContract.(*_AdsDiscoveryBlock).getLengthInBits(ctx))
 
 	// Const Field (statusLength)
 	lengthInBits += 16
@@ -239,7 +239,7 @@ func (m *_AdsDiscoveryBlockStatus) GetLengthInBits(ctx context.Context) uint16 {
 	return lengthInBits
 }
 
-func (m *_AdsDiscoveryBlockStatus) GetLengthInBytes(ctx context.Context) uint16 {
+func (m *_AdsDiscoveryBlockStatus) GetLengthInBytes(ctx context.Context) uint64 {
 	return m.GetLengthInBits(ctx) / 8
 }
